FIX NPE on enter world using auto supplies
