+++ Gameserver/src/main/org.l2j.gameserver/org/l2j/gameserver/engine/autoplay/AutoPlayEngine.java
@@ -323,7 +323,7 @@ public final class AutoPlayEngine {
             while (it.hasNext()) {
                 var player = it.next();
 
-                if(!canDoAutoAction(player)) {
+                if(player.getAutoPlaySettings() == null || !canDoAutoAction(player)) {
                     continue;
                 }
