+++ Gameserver/src/main/resources/sql/install/character_shortcuts.sql
@@ -3,9 +3,10 @@ CREATE TABLE `character_shortcuts` (
   `player_id` INT UNSIGNED NOT NULL DEFAULT 0,
   `client_id` INT NOT NULL DEFAULT 0,
   `type` ENUM('ITEM', 'SKILL', 'ACTION', 'MACRO', 'RECIPE', 'BOOKMARK') NOT NULL ,
-  `shortcut_id` decimal(16) ,
+  `shortcut_id` INT NOT NULL,
   `level` SMALLINT,
-  `sub_level` INT(4) NOT NULL DEFAULT '0',
+  `sub_level` INT NOT NULL DEFAULT 0,
+  `character_type` SMALLINT NOT NULL DEFAULT 1,
   `active` BOOLEAN NOT NULL DEFAULT false,
 
   PRIMARY KEY (`player_id`, `client_id`),
